disable system roots for windows
