+++ transport/internet/tls/config.go
@@ -3,7 +3,6 @@ package tls
 import (
 	"context"
 	"crypto/tls"
-	"crypto/x509"
 	"time"
 
 	"v2ray.com/core/common/net"
@@ -22,22 +21,6 @@ func ParseCertificate(c *cert.Certificate) *Certificate {
 		Certificate: certPEM,
 		Key:         keyPEM,
 	}
-}
-
-func (c *Config) GetCertPool() *x509.CertPool {
-	pool, err := x509.SystemCertPool()
-	if err != nil {
-		newError("failed to get system cert pool.").Base(err).WriteToLog()
-		return nil
-	}
-	if pool != nil {
-		for _, cert := range c.Certificate {
-			if cert.Usage == Certificate_AUTHORITY_VERIFY {
-				pool.AppendCertsFromPEM(cert.Certificate)
-			}
-		}
-	}
-	return pool
 }
 
 func (c *Config) BuildCertificates() []tls.Certificate {
