+++ transport/internet/tls/config_windows.go
@@ -1,0 +1,9 @@
+// +build windows
+
+package tls
+
+import "crypto/x509"
+
+func (c *Config) GetCertPool() *x509.CertPool {
+	return nil
+}
